+++ R/loadMantaNc.R
@@ -40,6 +40,14 @@ loadMantaNc <- function(x, keepQuals=c(1), keepEffort=TRUE) {
     on.exit(nc_close(nc))
     dimNames <- c('time', 'frequency')
     if(!all(dimNames %in% names(nc$dim))) {
+        # check for manta metadata NC
+        if(nc$ndims > 10 &&
+           nc$nvars > 10 &&
+           file.size(x) < 2e6) {
+            stop('NetCDF file format not recognized, it appears to be',
+                 ' a MANTA Metadata file that does not have SPL data.',
+                 'Try loading the MANTA CSV output instead.')
+        }
         stop('NetCDF file format not recognized, missing expected',
              ' "time" and "frequency" dimensions.')
     }
@@ -59,7 +67,7 @@ loadMantaNc <- function(x, keepQuals=c(1), keepEffort=TRUE) {
     # if qual flags present, replace some HMD with NA
     if('quality_flag' %in% names(nc$var)) {
         quality <- ncvar_get(nc, varid='quality_flag', start=c(1, 1), count=c(-1, -1))
-        qTypes <- unique(as.vector(quality))
+        qTypes <- sort(unique(as.vector(quality)))
         if(!all(keepQuals %in% c(1,2,3,4))) {
             warning('"keepQuals" expects values from (1, 2, 3, 4)')
             keepQuals <- keepQuals[keepQuals %in% c(1, 2, 3, 4)]
@@ -67,10 +75,13 @@ loadMantaNc <- function(x, keepQuals=c(1), keepEffort=TRUE) {
         # manuscript https://cdn.ioos.noaa.gov/media/2017/12/QARTOD_PassiveAcousticsManual_Final_V1.0_signed.pdf
         # these should be
         # "Pass", "Not Evaulated", "Suspect or High Interest", "Fail", "Missing Data"=9
-        dqLevels <- c('Good', 'Not evaluated/Unknown', 'Compromised/Questionable', 'Unusable/Bad')
+        dqLevels <- c('1 (Good)', 
+                      '2 (Not evaluated/Unknown)',
+                      '3 (Compromised/Questionable)', 
+                      '4 (Unusable/Bad)')
         dqDrop <- qTypes[!qTypes %in% keepQuals]
         if(length(dqDrop) > 0) {
-            message('Data quality flags ', paste0(dqLevels, collapse=', '),
+            message('Data quality flag(s) ', paste0(dqLevels[dqDrop], collapse=', '),
                     ' found in data, corresponding levels marked as NA.')
         }
         dropIx <- matrix(!quality %in% keepQuals, nrow=nrow(quality))
@@ -155,9 +166,11 @@ ncTimeToPosix <- function(vals, units) {
         units <- vals$units
         vals <- vals$vals
     }
-    # if(is.na(vals)) {
-    #     return(vals)
-    # }
+    
+    if(inherits(vals, 'POSIXct')) {
+        return(vals)
+    }
+
     isNa <- is.na(vals)
     
     if(grepl('hours? since', units, ignore.case=TRUE)) {
@@ -171,9 +184,19 @@ ncTimeToPosix <- function(vals, units) {
         }
         return(out)
     }
-    # if(units == 'hours since 2000-01-01 00:00:00') {
-    #     return(as.POSIXct(vals * 3600, origin = '2000-01-01 00:00:00', tz='UTC'))
-    # }
+    
+    if(grepl('minutes? since', units, ignore.case=TRUE)) {
+        or <- gsub('minutes? since ', '', units, ignore.case=TRUE)
+        or <- gsub('\\s{0,1}UTC', '', or)
+        or <- gsub('\\.0+$', '', or)
+        or <- ymd_hms_fast(or)
+        out <- as.POSIXct(vals * 60, origin=or, tz='UTC')
+        if(anyNA(out[!isNa])) {
+            warning('Conversion failed for units ', units)
+        }
+        return(out)
+    }
+
     if(grepl('seconds? since', units, ignore.case=TRUE)) {
         or <- gsub('seconds? since ', '', units, ignore.case=TRUE)
         or <- gsub('\\s{0,1}UTC', '', or)
@@ -185,21 +208,14 @@ ncTimeToPosix <- function(vals, units) {
         }
         return(out)
     }
-    # if(units == 'seconds since 1970-01-01T00:00:00Z') {
-    #     return( as.POSIXct(vals, origin = '1970-01-01 00:00:00', tz='UTC'))
-    # }
-    # if(units == 'seconds since 1981-01-01 00:00:00') {
-    #     return( as.POSIXct(vals, origin = '1981-01-01 00:00:00', tz='UTC'))
-    # }
+
     if(units == 'count') {
         return(vals)
     }
     if(units == 'posix') {
         return(vals)
     }
-    # if(units == 'hours since 1950-01-01') {
-    #     return(as.POSIXct(vals * 3600, origin = '1950-01-01', tz='UTC'))
-    # }
+
     stop('Dont know how to deal with time with units ', units)
 }