+++ R/loadSoundscapeData.R
@@ -28,6 +28,11 @@
 #'   \link{createOctaveLevel}
 #' @param label optional, if not \code{NULL} then this value will be
 #'   added as an additional column "label" to the output
+#' @param keepQuals quality flag values to keep. Accepts vector of
+#'   integers from (1, 2, 3, 4) corresponding to flag labels "Good",
+#'   "Not evaluated/Unknown", "Compromised/Questionable", and "Unusable/Bad".
+#'   HMD levels for points with data quality flags outside of \code{keepQuals}
+#'   will be marked as \code{NA}.
 #' @param keepEffort if \code{TRUE} or \code{FALSE}, a logical flag whether or
 #'   not to keep the effort information with the outputs (number of seconds
 #'   per minute). If a numeric value, then any minutes with an effort value
@@ -81,16 +86,17 @@
 #' @importFrom lubridate force_tz with_tz
 #' @importFrom future.apply future_lapply
 #'
-loadSoundscapeData <- function(x, 
-                               needCols=c('UTC'), 
+loadSoundscapeData <- function(x,
+                               needCols=c('UTC'),
                                skipCheck=FALSE,
-                               timeBin=NULL, 
-                               binFunction='median', 
+                               timeBin=NULL,
+                               binFunction='median',
                                binCount=FALSE,
                                octave=c('original', 'tol', 'ol'),
                                label=NULL,
+                               keepQuals=c(1),
                                keepEffort=TRUE,
-                               dropNonHmd=TRUE, 
+                               dropNonHmd=TRUE,
                                tz='UTC',
                                extension=c('nc', 'csv')) {
     if(is.character(x) &&
@@ -118,7 +124,8 @@ loadSoundscapeData <- function(x,
             loadSoundscapeData(f, needCols=needCols, skipCheck=skipCheck,
                                timeBin=timeBin, binFunction=binFunction,
                                binCount=binCount,
-                               octave=octave, label=label, keepEffort=keepEffort, 
+                               octave=octave, label=label,
+                               keepQuals=keepQuals, keepEffort=keepEffort,
                                dropNonHmd = FALSE,
                                tz=tz)
         }, future.seed=NULL))
@@ -132,15 +139,21 @@ loadSoundscapeData <- function(x,
             hmdLevels <- getHmdLevels(freqRange=range(freqVals)+c(-1, 1))
             nonStandard <- !standardHmd %in% hmdLevels$labels
             newLabs <- fixHmdLabels(freqVals[nonStandard], hmdLevels=hmdLevels)
+            repeatLabs <- newLabs[!is.na(newLabs)] %in% colnames(x)
+            if(any(repeatLabs)) {
+                warning('Input does not appear to be standard hybrid millidecade,',
+                        ' proceed with caution')
+                newLabs[!is.na(newLabs)][repeatLabs] <- NA
+            }
             colnames(x)[freqCols][nonStandard][!is.na(newLabs)] <- newLabs[!is.na(newLabs)]
             if(anyNA(newLabs) &&
                isTRUE(dropNonHmd)) {
                 warning('Found ', sum(is.na(newLabs)), ' non-standard ',
                         'hybrid millidecade frequencies (',
-                        paste0(standardHmd[nonStandard][is.na(newLabs)], collapse=', '),
+                        printN(standardHmd[nonStandard][is.na(newLabs)], collapse=', '),
                         ') these will be removed. Run with "dropNonHmd=FALSE"',
                         ' to keep them.')
-                for(col in standardHmd[nonStandard[is.na(newLabs)]]) {
+                for(col in standardHmd[nonStandard][is.na(newLabs)]) {
                     x[[col]] <- NULL
                 }
             }
@@ -163,12 +176,13 @@ loadSoundscapeData <- function(x,
             x <- fread(x, header=TRUE)
             setDF(x)
         } else if(grepl('nc$', x, ignore.case=TRUE)) {
-            x <- loadMantaNc(x, keepEffort=keepEffort)
+            x <- loadMantaNc(x, keepQuals=keepQuals, keepEffort=keepEffort)
         }
     }
+    colnames(x) <- checkTimeName(colnames(x))
+    x <- checkManta(x, keepEffort=keepEffort)
     if(isFALSE(skipCheck)) {
-        x <- checkTriton(x)
-        x <- checkManta(x)
+        # x <- checkTriton(x)
         x <- checkInfinite(x)
     }
     missingCols <- needCols[!needCols %in% colnames(x)]
@@ -188,26 +202,33 @@ loadSoundscapeData <- function(x,
         warning('Input "x" could not be formatted properly.')
         return(NULL)
     }
-    
+    # for now this check is just fixing 31_5 to 31.5
+    colnames(x) <- checkFreqNames(colnames(x))
     freqCols <- whichFreqCols(x)
-    freqVals <- colsToFreqs(colnames(x)[freqCols])
     type <- gsub('([A-z]*)_.*', '\\1', colnames(x)[freqCols][1])
     # standardizing to round to integer on all HMD columns
     if(type == 'HMD') {
+        freqVals <- colsToFreqs(colnames(x)[freqCols])
         standardHmd <- paste0('HMD_', round(freqVals, 0))
         colnames(x)[freqCols] <- standardHmd
         hmdLevels <- getHmdLevels(freqRange=range(freqVals)+c(-1, 1))
         nonStandard <- !standardHmd %in% hmdLevels$labels
         newLabs <- fixHmdLabels(freqVals[nonStandard], hmdLevels=hmdLevels)
+        repeatLabs <- newLabs[!is.na(newLabs)] %in% colnames(x)
+        if(any(repeatLabs)) {
+            warning('Input does not appear to be standard hybrid millidecade,',
+                    ' proceed with caution')
+            newLabs[!is.na(newLabs)][repeatLabs] <- NA
+        }
         colnames(x)[freqCols][nonStandard][!is.na(newLabs)] <- newLabs[!is.na(newLabs)]
         if(anyNA(newLabs) &&
            isTRUE(dropNonHmd)) {
             warning('Found ', sum(is.na(newLabs)), ' non-standard ',
                     'hybrid millidecade frequencies (',
-                    paste0(standardHmd[nonStandard][is.na(newLabs)], collapse=', '),
+                    printN(standardHmd[nonStandard][is.na(newLabs)], collapse=', '),
                     ') these will be removed. Run with "dropNonHmd=FALSE"',
                     ' to keep them.')
-            for(col in standardHmd[nonStandard[is.na(newLabs)]]) {
+            for(col in standardHmd[nonStandard][is.na(newLabs)]) {
                 x[[col]] <- NULL
             }
         }
@@ -267,11 +288,36 @@ checkTriton <- function(x) {
     if(alternate %in% colnames(x)) {
         colnames(x)[colnames(x) == alternate] <- 'UTC'
     }
+    alternate <- 'yyyy_mm_ddTHH_MM_SSZ'
+    if(alternate %in% colnames(x)) {
+        colnames(x)[colnames(x) == alternate] <- 'UTC'
+    }
+    x
+}
+
+checkTimeName <- function(x) {
+    if(is.data.frame(x)) {
+        names <- checkTimeName(names(x))
+        names(x) <- names
+        return(x)
+    }
+    tritonTime <- "yyyy-mm-ddTHH:MM:SSZ"
+    if(tritonTime %in% x) {
+        x[x == tritonTime] <- 'UTC'
+    }
+    alternate <- 'yyyy.mm.ddTHH.MM.SSZ'
+    if(alternate %in% x) {
+        x[x == alternate] <- 'UTC'
+    }
+    alternate <- 'yyyy_mm_ddTHH_MM_SSZ'
+    if(alternate %in% x) {
+        x[x == alternate] <- 'UTC'
+    }
     x
 }
 
 # colnames are d-m-y h:m:s, 0, 0-freq end
-checkManta <- function(x) {
+checkManta <- function(x, keepEffort=FALSE) {
     if(all(grepl('^X', colnames(x)))) {
         colnames(x) <- gsub('^X', '', colnames(x))
     }
@@ -289,14 +335,26 @@ checkManta <- function(x) {
     secondCol <- grepl('^0\\.{3}[0-9]{1}$', colnames(x)[2]) ||
         (colnames(x)[2] == '0' & colnames(x)[3] %in% c('0', '0.1'))
     checkSeconds <- checkSeconds & secondCol
-    if(isTRUE(checkSeconds )) {
-        x[[2]] <- NULL
-        colnames(x)[2] <- '0'
+    freqIx <- 2:ncol(x)
+    if(isTRUE(checkSeconds)) {
+        if(isFALSE(keepEffort)) {
+            x[[2]] <- NULL
+            colnames(x)[2] <- '0'
+            freqIx <- 2:ncol(x)
+        } else if(isTRUE(keepEffort)) {
+            colnames(x)[2:3] <- c('effortSeconds', '0')
+            freqIx <- 3:ncol(x)
+        }
     }
     # manta should have columns named just frequency for 2:ncol
     # if we cant convert w/o NA, then its not manta
-    freqCols <- colnames(x)[2:ncol(x)]
+    freqCols <- colnames(x)[freqIx]
     tryFreq <- suppressWarnings(as.numeric(freqCols))
+    # sometimes written as 31_5 instead of 31.5?
+    if(anyNA(tryFreq)) {
+        freqCols <- gsub('_', '.', freqCols)
+        tryFreq <- suppressWarnings(as.numeric(freqCols))
+    }
     if(anyNA(tryFreq)) {
         return(x)
     }
@@ -304,6 +362,15 @@ checkManta <- function(x) {
     if(is.character(x$UTC)) {
         x$UTC <- parse_date_time(x$UTC, orders=mantaFormat, tz='UTC', truncated=2)
     }
-    colnames(x)[2:ncol(x)] <- paste0('HMD_', colnames(x)[2:ncol(x)])
+    colnames(x)[freqIx] <- paste0('HMD_', freqCols)
+    x
+}
+
+checkFreqNames <- function(x) {
+    if(is.data.frame(x)) {
+        x <- colnames(x)
+    }
+    # grepl('[0-9]+_[0-9]+', x)
+    x <- gsub('(.*)([0-9]+)_([0-9]+)', '\\1\\2.\\3', x)
     x
 }